+++ src/unitexpr/__init__.py
@@ -17,5 +17,5 @@ The package `unitexpr` provides the modules:
 
 from .unit import UnitBase, UnitMeta
 from .unit_symbol import UnitSymbol
-from .qarray import QArray
+from .qarray import qarray
 from .quantity import Quantity